// mf-bank/src/app/models/ModelInput.ts
export interface ModelInput {

  CreditScore: number
  Age: number
  Tenure: number
  Balance: number
  NumOfProducts: number
  EstimatedSalary: number
  BalanceSalaryRatio: number
  TenureByAge: number
  CreditScoreGivenAge: number
  HasCrCard: number
  IsActiveMember: number
  Geography_Spain: number
  Geography_France: number
  Geography_Germany: number
  Gender_Female: number
  Gender_Male: number
}